Suggestions for the ECU Hard Reset
You don't need to send the same data with a different ID on this one. But my suggestions really, is to create options what are the UDS IDs on this one. 7DF is one of the standards and so is 7E1 but it could be different to other vehicles and not both of those IDs. You need to have an option to choose what ID and not just those two.